Fixed issue when switching application when application is starting
When an application A is starting, and an application B is stopped. If the user select the application B in the application list, then the "current_application" for the application_view_component becomes the application B. So after application A has finished starting, the application B in the application_list_component is overwritten by the data of the application A. This commit fixes this issue by saving the application A in a local variable.