Fixed 404 error when clicking on a stopping application
When clicking on a stopping application, Ember was trying to render it resulting in a 404 error. It's fixed by rendering the iframe only if the application is running, the form will be rendered for every other application status and its buttons will be disabled if the application is not completely stopped.