Fix status listening issue
Using a Vue object listening to the current_app.status variable results in issues when current_app is changing. Because if current_app changes (when clicking on a new application) and the new clicked application does not have the same status as the previous one, the event is triggered. We now use the emit method to solve this issue