Fix loading issue
Vue was complaining that it couldn't find the dom elements with id #appview and #applist (containers for the Vue instances).
I assume that, before using webpack, it was taking more time for the javascript to be loaded. So when the javascript code was starting to be executed, the whole page (including #appview and #applist) was already mounted. But now that the javascript file is minified, it takes less time to load it and execute it, and the dom elements are not ready yet. So we need to load the javascript only when the page is mounted. Meaning that the script element containing the javascript must be the last dom element of the body.