Fix close button issue
As it's not a "button" element but an "li" element, there is no disable attribute, we only play with the disabled class which change the visual effect but doesn't prevent the click event to be triggered.
So the click event is still triggered and we have to deal with the case when the application is not currently running.