Fix an other id issue
When changing the url from an existing id "id1" to an unexisting id "id2", an issue is displayed but the table of the id "id1" is still displayed. In order to fix that we have to empty the table before trying to retrieve informations about the new id, so that if the new id is unknown, the table is empty as expected.